# backend-nodejs/docs/api-documentation.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Business Intelligence API for University Admission Data",
    "version": "1.0.0",
    "description": "API for processing and analyzing university admission test data",
    "contact": {
      "name": "API Support",
      "email": "support@university.edu"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AdmissionStats": {
        "type": "object",
        "properties": {
          "totalStudents": {
            "type": "number",
            "description": "Total number of students"
          },
          "averageScore": {
            "type": "number",
            "description": "Average score across all students"
          },
          "maxScore": {
            "type": "number",
            "description": "Highest score achieved"
          },
          "minScore": {
            "type": "number",
            "description": "Lowest score achieved"
          },
          "approvedCount": {
            "type": "number",
            "description": "Number of students who were approved"
          },
          "rejectedCount": {
            "type": "number",
            "description": "Number of students who were rejected"
          },
          "pendingCount": {
            "type": "number",
            "description": "Number of students with pending status"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
      "DemographicBreakdown": {
        "type": "object",
        "properties": {
          "gender": {
            "type": "object",
            "properties": {
              "male": {
                "type": "number"
              },
              "female": {
                "type": "number"
              },
              "other": {
                "type": "number"
              }
            }
          },
          "residence": {
            "type": "object",
            "properties": {
              "provinces": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "count": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "educationType": {
            "type": "object",
            "properties": {},
            "additionalProperties": {
              "type": "number"
            }
          }
        }
      },
      "ComponentPerformance": {
        "type": "object",
        "properties": {
          "components": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "averageScore": {
                  "type": "number"
                },
                "maxScore": {
                  "type": "number"
                },
                "minScore": {
                  "type": "number"
                },
                "passRate": {
                  "type": "number"
                }
              }
            }
          }
        }
      },
      "DisabilityStatistics": {
        "type": "object",
        "properties": {
          "withDisability": {
            "type": "number",
            "description": "Number of students with disabilities"
          },
          "withoutDisability": {
            "type": "number",
            "description": "Number of students without disabilities"
          },
          "disabilityTypes": {
            "type": "object",
            "properties": {},
            "additionalProperties": {
              "type": "number"
            }
          }
        }
      },
      "AdmissionTrends": {
        "type": "object",
        "properties": {
          "periods": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "period": {
                  "type": "string"
                },
                "year": {
                  "type": "number"
                },
                "totalApplicants": {
                  "type": "number"
                },
                "approvedCount": {
                  "type": "number"
                },
                "rejectedCount": {
                  "type": "number"
                },
                "averageScore": {
                  "type": "number"
                }
              }
            }
          }
        }
      },
      "DashboardData": {
        "type": "object",
        "properties": {
          "statistics": {
            "$ref": "#/components/schemas/AdmissionStats"
          },
          "trends": {
            "$ref": "#/components/schemas/AdmissionTrends"
          },
          "demographics": {
            "$ref": "#/components/schemas/DemographicBreakdown"
          },
          "performance": {
            "$ref": "#/components/schemas/ComponentPerformance"
          },
          "disability": {
            "$ref": "#/components/schemas/DisabilityStatistics"
          }
        }
      }
    }
  },
  "paths": {
    "/api/bi/statistics": {
      "get": {
        "summary": "Get admission statistics by period and year",
        "description": "Returns statistical data about admissions for a specific period and year",
        "tags": [
          "Business Intelligence"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "schema": {
              "type": "string"
            },
            "description": "Academic period (e.g., 20252)"
          },
          {
            "in": "query",
            "name": "year",
            "schema": {
              "type": "number"
            },
            "description": "Academic year (e.g., 2025)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionStats"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/bi/trends": {
      "get": {
        "summary": "Get admission trends over time",
        "description": "Returns admission data trends across multiple periods",
        "tags": [
          "Business Intelligence"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionTrends"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/bi/demographics": {
      "get": {
        "summary": "Get demographic breakdown of applicants",
        "description": "Returns demographic data about applicants including gender, residence, and education type",
        "tags": [
          "Business Intelligence"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemographicBreakdown"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/bi/performance": {
      "get": {
        "summary": "Get performance analysis by component",
        "description": "Returns performance data broken down by test components",
        "tags": [
          "Business Intelligence"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComponentPerformance"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/bi/disability": {
      "get": {
        "summary": "Get disability statistics",
        "description": "Returns statistics about applicants with disabilities",
        "tags": [
          "Business Intelligence"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisabilityStatistics"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/bi/periods-years": {
      "get": {
        "summary": "Get available periods and years",
        "description": "Returns a list of available periods and years from admission data",
        "tags": [
          "Business Intelligence"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "periods": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "years": {
                          "type": "array",
                          "items": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/bi/dashboard": {
      "get": {
        "summary": "Get comprehensive dashboard data",
        "description": "Returns all necessary data for the business intelligence dashboard",
        "tags": [
          "Business Intelligence"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "schema": {
              "type": "string"
            },
            "description": "Academic period (e.g., 20252)"
          },
          {
            "in": "query",
            "name": "year",
            "schema": {
              "type": "number"
            },
            "description": "Academic year (e.g., 2025)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardData"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}